Re-organize the symmetry-tools module to be more extendable
