+++ symmetry-tools/pom.xml
@@ -114,7 +114,7 @@
 										<Implementation-Version>${project.version}</Implementation-Version>
 										<Specification-Title>${project.name}</Specification-Title>
 										<Implementation-Title>${project.name}</Implementation-Title>
-										<Main-Class>demo.CeSymmMain</Main-Class>
+										<Main-Class>main.CeSymmMain</Main-Class>
 									</manifestEntries>
 								</transformer>
 								<!-- This bit merges the various META-INF/services files -->
